Move functions to new submodule
